shuffle around and rename site files
